update project fatal: The upstream branch of your current branch does not match the name of your current branch.  To push to the upstream branch on the remote, use
    git push origin HEAD:project-2-builer-button

To push to the branch of the same name on the remote, use

    git push origin HEAD

To choose either option permanently, see push.default in 'git help config'.

To avoid automatically configuring an upstream branch when its name
won't match the local branch, see option 'simple' of branch.autoSetupMerge
in 'git help config'.